fix: pass personalization to workflow runs
